+++ application/service/git/git_test.go
@@ -3,7 +3,7 @@ package git_test
 import (
 	"fmt"
 	"github.com/duck8823/duci/application/context"
-	"github.com/duck8823/duci/infrastructure/git"
+	"github.com/duck8823/duci/application/service/git"
 	"github.com/google/uuid"
 	"gopkg.in/src-d/go-git.v4/plumbing"
 	"net/url"
