+++ application/service/git/progress_logger.go
